rev: change trailing return types to traditional; further separate forcings engine impl
