+++ src/forcing/CMakeLists.txt
@@ -27,7 +27,11 @@ if(NGEN_WITH_NETCDF)
 endif()
 
 if(NGEN_WITH_PYTHON)
-    target_sources(forcing PRIVATE "${CMAKE_CURRENT_LIST_DIR}/ForcingsEngineDataProvider.cpp")
+    target_sources(forcing
+      PRIVATE
+        "${CMAKE_CURRENT_LIST_DIR}/ForcingsEngine.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/ForcingsEngineDataProvider.cpp"
+    )
     target_link_libraries(forcing PUBLIC pybind11::embed NGen::ngen_bmi)
 endif()
